refactor code, add controllers for cart and wishlist route
