+++ index.js
@@ -1,12 +1,12 @@
 const express = require("express");
-const bodyParser = require("body-parser");
 const cors = require("cors");
 require("dotenv").config();
 
 const app = express();
-app.use(bodyParser.json({ urlExtended: true }));
+app.use(express.json());
 app.use(cors());
 const { initializeDBConnection } = require("./db/db.connect.js");
+
 initializeDBConnection();
 const productRouter = require("./routes/product.router.js");
 const userRouter = require("./routes/user.router");
@@ -15,6 +15,7 @@ const wishlistRouter = require("./routes/wishlist.router");
 const paymentRouter = require("./routes/payment.router");
 
 const authVerify = require("./middleware/authVerify.js");
+const routeHandler = require("./middleware/routeHandler");
 
 app.use("/user", userRouter);
 app.use("/products", productRouter);
@@ -26,13 +27,7 @@ app.get("/", (req, res) => {
   res.json({ success: true, message: "Ecom SeverV1" });
 });
 
-// 404
-app.use((req, res) => {
-  res.status(404).json({
-    success: false,
-    message: "route not found on server, please check",
-  });
-});
+app.use(routeHandler);
 
 app.listen(process.env.PORT || 7000, () => {
   console.log("Server Started");
