+++ middleware/routeHandler.js
@@ -1,0 +1,9 @@
+/* 404 Route Handler */
+const routeHandler = (req, res) => {
+  res.status(404).json({
+    success: false,
+    message: "route not found on server",
+  });
+};
+
+module.exports = routeHandler;
