+++ models/cart.model.js
@@ -1,7 +1,7 @@
 const mongoose = require("mongoose");
-const { Schema } = require("mongoose");
 const { User } = require("../models/user.model");
 const { Product } = require("../models/product.model");
+
 const CartSchema = new mongoose.Schema(
   {
     productId: { type: mongoose.Types.ObjectId, ref: Product },
